requirements.txt
# Core dependencies
agency-swarm>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Blockchain dependencies
web3>=6.11.1
solana>=0.25.6
solders>=0.19.0
anchorpy>=0.19.0
base58>=2.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
textblob>=0.15.3

# API and networking
aiohttp>=3.8.0
websockets>=10.0
tweepy>=4.14.0
requests>=2.28.0

# Database (optional, based on your storage needs)
# sqlalchemy>=2.0.0
# pymongo>=4.3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0